src/generator/requirements.txt
faker==25.0.0
psycopg2-binary==2.9.9
